Add constant groups post

+++ src/routes/tracer/rules.rs
@@ -2,7 +2,7 @@ use rocket::{response::Redirect, serde::json::Json};
 use rocket_okapi::openapi;
 use rocket_sync_db_pools::diesel::prelude::*;
 
-use crate::{db, models::rule::Rule, schema::rules};
+use crate::{db, models::rule::Rule, schema::tracer::rules};
 
 /// Lists all stored rules in the database
 #[openapi(tag = "Rules")]
@@ -17,7 +17,7 @@ pub async fn get_rules(conn: db::DbConn) -> Result<Json<Vec<Rule>>, Json<&'stati
 /// Lists all stored rules in the database
 #[openapi(tag = "Rules")]
 #[get("/rules/<id>")]
-pub async fn get_rule_id(id: i64, conn: db::DbConn) -> Option<Json<Rule>> {
+pub async fn get_rule_id(id: i32, conn: db::DbConn) -> Option<Json<Rule>> {
     let rule: Rule = conn
         .run(move |c| rules::table.find(id).first(c))
         .await
@@ -29,7 +29,7 @@ pub async fn get_rule_id(id: i64, conn: db::DbConn) -> Option<Json<Rule>> {
 /// Lists all stored rules in the database
 #[openapi(skip)]
 #[get("/rules/<id>/go")]
-pub async fn redirect_to_rule(id: i64, conn: db::DbConn) -> Option<Redirect> {
+pub async fn redirect_to_rule(id: i32, conn: db::DbConn) -> Option<Redirect> {
     let rule: Rule = conn
         .run(move |c| rules::table.find(id).first(c))
         .await
